#6 and correction to `create_dataset_from_slim_output()`
While working on #6, I found out that there was a bug in `create_dataset_from_slim_output()` which misattributed samples to the wrong deme as the deme ID was changed during the tree simplification step. Corrected this and updated the selective sweep figure in devlog-20250113. This made the figure noisier but did not change the overall message about selection affecting effective migration rate estimates.